pyproject.toml
[project]
name = "py_load_spl"
version = "0.1.0"
description = "A Python package for loading FDA SPL data into a database."
authors = [
    {name = "OHDSI"},
]
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "lxml>=6.0.1",
    "requests>=2.32.5",
    "tenacity",
    "pydantic>=2.11.7",
    "pydantic-settings",
    "typer>=0.17.4",
    "beautifulsoup4>=4.13.5",
    "python-json-logger>=2.0.7",
    "pyarrow>=21.0.0",
    "xmltodict>=0.15.1",
    "boto3>=1.34.0",
    "databricks-sql-connector",
]
readme = "README.md"

[project.optional-dependencies]
postgresql = ["psycopg2-binary"]
redshift = ["redshift-connector"]

[project.scripts]
py-load-spl = "py_load_spl.cli:app"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true

[[tool.mypy.overrides]]
module = [
    "pyarrow.*",
    "redshift_connector.*",
    "testcontainers.*",
    "pythonjsonlogger.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src/py_load_spl --cov-report=term-missing --cov-fail-under=94"
testpaths = [
    "tests",
]
pythonpath = [
    "src"
]
markers = [
    "integration: marks tests as integration tests that require external services",
]
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated:DeprecationWarning",
]


[tool.pdm]
distribution = false
[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov",
    "testcontainers",
    "psycopg2-binary",
    "ruff>=0.12.12",
    "black>=25.1.0",
    "mypy>=1.17.1",
    "pre-commit",
    "requests-mock>=1.12.1",
    "python-json-logger>=2.0.7",
    "pytest-mock>=3.12.0",
    "moto[s3]>=5.0.0",
    "redshift-connector>=2.0.910",
    "types-requests>=2.32.4.20250809",
    "types-psycopg2>=2.9.21.20250809",
    "types-xmltodict>=0.15.0.20250907",
    "boto3-stubs[s3]>=1.40.27",
    "lxml-stubs>=0.5.1",
    "pandas-stubs>=2.3.2.250827",
    "memory-profiler>=0.61.0",
]
